@@ -2,14 +2,13 @@ import React from "react";
 import TouristInfoCards from "./components/TouristInfoCards";
 import Bookings from "./components/Bookings";
 import Heading from "./components/Heading";
-import SearchResults from "./components/SearchResults";
 import Footer from "./components/Footer";
 import "./App.css";
 
 const address = [
   "123 Fake Street, London, E1 4UD",
   "hello@fakehotel.com",
-  "0123 456789"
+  "0123 456789",
 ];
 
 const App = () => {
@@ -18,7 +17,6 @@ const App = () => {
       <Heading />
       <TouristInfoCards />
       <Bookings />
-      <SearchResults />
       <Footer footerAddress={address} />
     </div>
   );
