@@ -4,7 +4,10 @@ const Heading = () => {
   return (
     <header className="App-header">
       CYF Hotel
-      <img src="/week-1/cyf-hotel-react-1/hotel.png" />
+      <img
+        className="hotel-logo"
+        src="data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wCEAAkGBw4QEBASExIWFRQWGRgTFhcXFRkQGBkbHRUYGRgcGBUYHSggJBolGxYWITEhJyorLi4vGR8zODUtNygtOisBCgoKDg0OGhAQGzUmICUtLS0vLSsvNS8wLS0tLSstLS8tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLS0tLf/AABEIAMIBAwMBIgACEQEDEQH/xAAcAAEAAgMBAQEAAAAAAAAAAAAABQYDBAcCAQj/xABHEAABAwIEAwQECQoEBwEAAAABAAIDBBEFEiExBhNBIlFhcRQjMoFCUlRicpGhscEHFRYkMzRzkpOzQ9HS8DVEgoOy0+El/8QAGgEBAAMBAQEAAAAAAAAAAAAAAAIDBAEFBv/EADQRAAIBAgMFBgUDBQEAAAAAAAABAgMREiExBEFRYfATcYGxweEFMpGh0SIj8RQVM3KyBv/aAAwDAQACEQMRAD8A7iiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAi0sSxKCmYZJpGsaOpP3DcnwCxYVjVJVgmCZkltwDqPNp1C7Z2vbI5iV7bzYra6GFuaWRrB3uICywzNe1rmm7XC4I2IXLcED8RxKaok1ijcXauyta1psweVtTt4kXXR6XFqaV2RkrXEdAfuOxU6lPBlv38iEJ4lckERamJVbYIZZnAlsbHSOA3s1pcbeOirLGzaUdR43SzSPijma97NHAG/nY7G3W2y5/iFXiOJMMk0n5vw/qXGz5Bf3ON+7Qa/CUBQNoM0/IdIxoc30d7rgmzBmLu679RtYdy2U9kxJpvPlnbvengY6m14bNLLnlfu3+J3FFRsF4jqopWU9Q3mtc5rGzN73Wy5uh0I8fNXlZ6lNweZop1VUV0ERFWWBERAEREAREQBERAEREAREQBERAEREAREQBERAERfDsgOa/lFqMJbN+sCeWfKMrGOytY3zNgATqbXJ8lzyGtZBKyWldKx7Te7sv1C3TpY3XTaLgBs73VNa9z5ZSHuZewZqTkBHcMrfIG24tDcR4DDU4myipY2xBjRJO8N1FwCB5BuUADq5enRq04rBdvLPh9Dza1ObeOyV3lxKrQVLg1zL2a4gutoTbYHwGpt3q8cOPogQHNkPcdBbzAP3arRdwqzkVTobmSnqHw2JvmZ2Mt+lxmBv5qTw2tohTxmSP19yxjItXyf9PTU9e4+ShVl2nyl9NYPmZ0KEjK2xuLaE9VHcVfuNb/Am/tuVNn4pxCmnymGMxhrS6nD7ytDnWFn9X7XHiNlN1fENNXYdXOhfq2GUPY4ZXsPLdo5v167aLK6MoNN6ZaFirRmmlrnqVTEGxMlY6vkdX1h/Z0cOsbNNA5oHd4e47rW5lXnrjPTsBLo+axn+F6kcvKL20Zv49QtzB3OLXMwuIRx6iWvn3Pxsubxv/kN1FUUcYfViKsdKQ5uV7ibT+rHMvf5+g1Onet9PV3/jNbllFcs3veRhq6K2/wC+XF5y8tyzJnBW+tYYJLwc2HMx3tt7ce3kbNt4dV1Fcro3MNZFzWcufmxWy+w4ZmA7d+rvxKt+OcWRQSejwsdU1R2hj6eMjtmjVZ9pi5SVlu6z0feX7LNRi78estV3FlRc/wAL4srWPk9JZHI0SFjhEbmMhodp8ZoF/e06q54biMNRGJInhzTpcdD3EdDrss1SlKGb04mmFaM8lqbqIirLQiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIAiKpcYcUy0jmxU8POmIzuGuVjdQC63eRoLjZShBzdkRnNRV2a9Tjr3Yv6O52SGnidM7UgOJYDmefitBOnf9mpwXjEc9RVTR07zzZDnncWsYGjSMNDtT2WtuN1zniHHZquodIWNZI5ggfynFzX66W1Ovwdzey3uF8JirJhT1E7oXggMjyZL6XIAIyh/Xa5XovZlGneWWS5+OWufgectocqllnm+XcuVvrwOn4CAyvxSE7PdHOB3h0YaftaoN8RppZW83lZj/AIbOZO+7ndiLuF9z4LMyjfh+I0ZdM+VkzDT5n6uHaBaCeozFtr62W9xLLkqXHmtiBjY3M1vMndd0nYhHS/U+Szw+fLejTP5M9z63lIxyjyz29FyBrY3WbJnnZeQ9u/WQ31b4tUjw2/NSYqczH/q0naAyyH9tpK342v2qOx2lyz/uzo8rY3XEmaoZeQ9vxkPVv0VI8NuJpMVdmjfenks9oyvP7b9q3o/X7Vpqu9O/d5rm/My01af18vA+4u28bH4rMIYQAYqGD2nAbZre7/MKDo8SpZTUkU4jje5vLa035WVgaLGw39o+PQqpVE75HF73Oe47ucS4nzJUphHsHz/ALVToYdX9MkuuLzM1Ss5aL65t9/4L5hXMEsYa4TQ86H1hPaabx2+oZW9/lssmLyXqamIyOIklP6vTt/WJ7Mb+1l+BEPuuoDhuVwqqdocQHSR3F9D22nUeYCnsYktVVMZkIEkluTSt/Wp7NbpJJ8CId99RmWWtDDU8OuvLU1UJ4oePXX2ehD08PLMl4Ww5ZHt9Q7mmL1Y7Hzm9Se8PVu4aqnRwl92G8pvy9GuHKj1+l+KqNPDyzJ6nkZZXN9U7m8v1YOQ/Gb1J+krNg/7s72NZSbx+yfVR6+Z6+K8v4/UlT2CU4PNONn4rrUv+HRUtpSfMvUcgcA4G4IuFkVd4Wrb8yE/B7TfInX7fvViVOz1u2pqfHpno1qTpTcWERFcVBERAEREAREQBERAEREAREQBERAEREAVG/KFQ1FU+mpIDl5xe+YgWuxgaO2eoBeOz4hXlYywXBtqLgHrrv9w+pTpzwSUluIVIY44TnkXBdLFiFPCI7xGnlLidSX3DCSfJ+nctxuHCrD6afs11MWuZNaxka03hlB63Gju438FZ8dp80E7m3bKIpGseDZwuL6Hza0+5UPBsb/O0UYziHEoAXRP0aJR8IW6ggdpvTcaK+MpzWJvTfw59z0ZnlGMHhS13ceXetxZPyh0hfRtkGj4ntkuOnwT94PuVc4g4ifmiliDWySQMvJa722dIHNaTtc31ViwzHGV8E9NM3lVIY9kkR01y2LmX3HXw+onnVb7MH0T/AHHq3ZYXnhktL/dFe1T/AG8UXrYhMWneJGPDjn1dmub3uDe/mrXwZWmajxgua0OFM7M4DKX3bMbu8dSqhjPtM8j94WnFVSMY9jXua19s4BsHWvbN3jU6L0a1PHG3cefRqYJXMSl8I9g+f4BQ6mMI9g+f4BWlZPcO/vdL/Ej/APML7xtj08NZWQQ2izPvJIwZZXgsbZpk3DR3DvUcCtXFad8z3SFxLza5drewA377AKqVJSmpPcvW5OFXDBxXH0se8DnfGwOYS03Oo/34roHD82ekc7K1vrXXDRYX5UetvFc8w5hbGARY3P3q/wDCn7kf4r/7ca8P/wBSl/bZvnH/AKRv+Dt/1cV3+RtYLMWVsXzszT/IT94Cvy5/g8ZdXQ+Bc4+5jvxsugLyvhf+DxPo/idu0j/r6sIiL0TzQiIgCIiAIiIAiIgCIiAIiIAiIgCIiAIiIDy4A6FcAruGq+mqpBFFL6uQ8t7AdgbtII8LL9ArgH5QeV+cq65dnztsLDLblR7m9779Fu2FyxNLh1w4mHbksKb4lkxKSWrhjllpZo61nYc5kRySt7yRs7X7xta0PiNNWObGI6WW7WkElhAHbcfuKqjuR2LF3z7tbptfJ2teu9ui9N9HzOu5+T4JDWl1/EZrW36lboUXB3Xl7mKdfGrPr7G+/AMQcbmnlJ+iV8/R2v8Ak0v8hUcOVkNyc99BYZbeLr3v7l9dyOxZzraZ7tbptfJ2teu9lbafS9yr9HT9iQ/R2v8Ak0v8hW/QYTWMaQaabe/sE9ygm+j5nXc/J8Ehrc1/nDNYDyJWMcrIdXZ76Cwy28XXvffolp9L3H6On7Fr/N9V8mm/plPzfVfJpv6ZVXdyLss59vh3a3TvydrXrvZG+j5nXc7J8Ehrc1/nDNYe4lP3Ol7nLQ6fsWj831Xyab+mVY8CqZIaYxvp6jNzHP0iJFixg382lcxHKybnPfawy2+le9/cvTuRdlnPtpn7LbjvydrXrvZZdu2NbZRdGr8rtouGfEu2ev2FRVIa8/4O28GsL5JZTG9mUBoztyXzam38o+tXFcy/IxktXZCS3NFa4APsv3AJC6avI/pobM+yhoj2ltE9oSqT1foEREOhERAEREAREQBERAEREAREQBaNHiUEr5o2Ou6EhsgsRlJFwNfwW8ucYXNiDa/FfRIopAZW5+Y8x2OU2tYeashDEny/JXOeG3P8MvEGKQPmmha8GSHKZBYjKHC4udtlDycd4UH5PSRobFwa4sv/ABAMv2qi181Z/wDvue1rJjHAJGxuLmhpDQ6zj0yb+ZXRcMoqL0GNjWxmmMYJuBlLctyXHv6klWTpxgryz007k+HPIrhUlPJZa+bXofMT4sw+mc1ss7WlzQ9tgXgtOxBaCLaLDTcbYZLnyTg5GOld2XaNba528Qq3wyC7AagnVobUNiLtTkBdaxPS+ZbpY39Hb2F/RBrbX2B1XXSgv0u97219LBVJtKStpfrMkHflCwcf80P5X/6Vjmwumqqupa2rnErMpkjbkDWZmjLYujO413O6jeHhi/otNkpaQx8uPK5znZy3ILFwt7VlucN/8axnyp/7QRwjG7i9FxT3pcDilKWFSWvJrc3xNz9DY/lNT9cP/qUbiGE0NPJFHLXTsfKbMBMeuoHSLTUjdXkrjOL4vh1dPiUlROGHLyKQZXutlJOe7QbZnDzsSlGMqjd9FwXt1YVnGmlZZvr28S54rw/SUsZlmrJ2MBALvVbk2G0S8YhgdHTw8+StnEXZ7QMTgc1g22WI6G4UNjuM+nYC2V2rw+OOT6QeAT7xY+9RvFpnoaaTD5Mz4XuZJSyHWzRI0uid9Ebf/dLKdKTaTed7PTRW0y3XK51IxTaWVrrXV8SbxJuF00ropcQqGPba4sw7gEaiHuIUlhmA0dSzmQ108jdiWuiNj3Ecq4PmmBRtdjOJ3APq6fcX+AFjwSOKPGqtlOAIzC10zW6NEuYW0GgOUn7VXK1stUk91t3LmTis89G7b91+Zp4rFhdLKYpsQnY8AEghh0IuNRDZZIKTDZIZJ2YhO6OO2cgxktvtdvKv9ixYhXywYzVvjpXVJ5EeZrS1paNDm7W/dYaqHkzVVLi2IhjIopYmxNja4ON2vbmc8ACzvt1+uxU8k3y4PXda2XLuIOVm1bjx3c+tSQZVYKSAMSnuSANGbn/sraxenw6kk5UldU8y2YsY1krgO8hkJt71t4QMY5dP+q0nLyx6lzs2Ww1tb2rL3U4RVsrampoZ6dzpMjZopQXZS0WFnMNxcdNPeotxva/3Xmo5epLDK2n2f5zJHhN9I2CSaGqdNGdXOkLBkyg3BAa3Lob6r5Fx3hTnhgqBqbBxa5rCfpkZftVQ4gxh01BUwyQtp3RVETKvleyWvfcuBHeRrfXbvV3xqjo/QJmuawU4icRYCwAbdpb47EEdbKEoRTvK+b3eHKz15fcnGUnlG2S/P005mbGeI6KjLBPKGZwS3QuuBa+oB7wvOF8UUFVmEM7XloLiNQ6w3IaRcjyXO4KqoaeH5OU6eQRz5Y8waXtykN7TtNGWOvcpfBJ5MQxOOcQNphSB8crS4GVznAgBzQB2RrqdN/dKWzqMc+ed1ubWmuZFV5OXisrPgnrpvJ4flBwj5SP5H/6VO4biEVTE2WJ2Zjr5XWIvY22PiFV8cib+ecNFhblzaWHcrkGgbafYqqiiksKeavrzfLkW03Jtp7stPc9IiKotCIiAIiIAiIgCIiAKOw/CIYJaiVl807g+S5uLgWFh03Uii7dnGkyLiwOnbNUz5SXVAa2UOOZpDW5QMp02VbxHgfDomOdlqTFe5gikkc03PxAb296vC8u2Kkqk46NkHSg9UQWEzUtbRmOOKSOncwMbdhhBY5vwL9LdVgwR1BW0D6eFzjTsBpje7HZQ0WNzrYtIId1BusrMGnZhjKNkrWyCJsJkANgNA9zRvfLe3jZYMF4YkpZiW1DpIXwiF7HtYwjJYRFnLa0aNLwbi+re5Ru/UnhXAruHcK4PUODY46zluzBkt5GxOy3vld8XQ2cQAehNwpmXhfDq6oqZrzCQScuXLI6IZmsaNAOmXL9q+4bw1Vxy0YfLEYqXsMc3mNkkjDHMYyRmbJoHAl2ty0aBb+H4fWw1E5DoDBLMZiCH8wAtaLA3y37Kn21TXEyHY09MKNKE4fS0NawPmEML3xTOJc94LmszZHb2tI3UeK3cHkoIGejQgFsMLKi9s4MbzJlcH9STG9eJeH5DDXR523qJxO02Nmi0Oh8fVH6wvGH8MCCetfG/1c8TY2Rn/CIMrnBvzLy3Delz0soOTerJKKWiI7D8Mw6tE4bDUsjqCyocXMdDGSDdpYTsTe9hv7lKcYMo5Io6eojfKZHerZE0uku0XLm22AG7iQNbdVh4Q4enoyA/kWEYjvHzsxItqRI8ttodgFuY7hU8ksM8DmCRjZIi2UOyPjkylwzMOZrrsYQ7XqLaruOWWegwRzy1IPGsDwyUPrpxURl+Vhb6yN5IIja0RAXJJAAte+40Unwq2gp2SxU8L4XMAfIyRrmyEEGzruJzN0IBBIFiNF4h4cnbSMi5rRMyb0ljrPfG13MLgyznZiwNJbvfrotqiwyqdJNPUOj5j4hAxsQdka27jcudqXFzu4WA9666kmsLeRxU4p4kszW4fqKSpqTWRRVAdNC20kkbo43R9kty30uQQfK60aKhw+rfUiOKobHVBwe4NcyCQsdq5u+V1wdSG31tdbHBnDlRRZGyckhsQizM52dxGXUh7ywXy3NgPDRMM4cqY6uOYmCNrXSOfyA+Mz5muAEkROQWJDiRckgba34pyWjOuEXqjTquCMJp3QF3Pu+RsTLSucMxuRfw7JWTH8Iw1z62qLp4HwNz1D4XujLhy8+w3OXuU9xFhHpbYGEjKyZkrhci7WhwIBbqD2go6XhS0GJwxODG1TCxly5+QmHlkuLiSddVLtqjd3JkOxp2soox4JSUEdLUximmbDlMk3PjcDJoSTd2pNh020UNhPD2ETyMitVhpbz44J3Ssjc0EXLAdwC5txe/aHep3CuHp44auN3IbzY+W0x8465Xjtc17tO108V7wPho0s7ZOYZByRF6x75XRkZc3KLibRuyglve0eSKrNNtN5nXSg0k4rIx0M1HWVNLMyGoBha8QvMbo4rEFh12OmgUpJgMDqptWA5swGQua7KHt7nt2P8AvuCh+F+GqikkaX8gtAcMzedzDc3Gjn5PqCt6jilx5eBLBHhz8SNqcHhkqYalwPMiDmssbCzt7jqpJEXG29TqSQREXDoREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQBERAEREAREQH/9k="
+      />
     </header>
   );
 };
